Abstract HTTP handler logic to Http class
